Create 2010census_popByZip
This code makes a csv file with total 2010 population by state and zipcode. You will need to get an api key and put it in as well as update the path for the csv file.

Slight modifications will pull data for different geographic regions or different variables. See the following for information on how to change the url to reflect these changes: http://api.census.gov/data/2010/sf1/examples.html

And see the following for variable names:
http://api.census.gov/data/2010/sf1/variables.html